requirements.txt
Flask==2.3.2
gunicorn
flask-sqlalchemy
flask-migrate
psycopg2
boto3
python-dotenv